+++ packages/sst/src/constructs/NextjsSite.ts
@@ -85,7 +85,7 @@ export class NextjsSite extends SsrSite {
 
   constructor(scope: Construct, id: string, props?: NextjsSiteProps) {
     super(scope, id, {
-      buildCommand: "npx --yes open-next@2.0.4 build",
+      buildCommand: "npx --yes open-next@2.0.5 build",
       ...props,
     });