+++ packages/sst/src/constructs/SsrSite.ts
@@ -69,6 +69,7 @@ import { Secret } from "./Secret.js";
 import { SsrFunction } from "./SsrFunction.js";
 import { EdgeFunction } from "./EdgeFunction.js";
 import {
+  BaseSiteFileOptions,
   BaseSiteReplaceProps,
   BaseSiteCdkDistributionProps,
   getBuildCmdEnvironment,
@@ -98,6 +99,7 @@ export type SsrBuildConfig = {
 
 export interface SsrSiteNodeJSProps extends NodeJSProps {}
 export interface SsrDomainProps extends DistributionDomainProps {}
+export interface SsrSiteFileOptions extends BaseSiteFileOptions {}
 export interface SsrSiteReplaceProps extends BaseSiteReplaceProps {}
 export interface SsrCdkDistributionProps extends BaseSiteCdkDistributionProps {}
 export interface SsrSiteProps {
@@ -270,6 +272,60 @@ export interface SsrSiteProps {
       | "logRetention"
     >;
   };
+  /**
+   * Pass in a list of file options to customize cache control and content type specific files.
+   *
+   * @default
+   * Versioned files cached for 1 year at the CDN and brower level.
+   * Unversioned files cached for 1 year at the CDN level, but not at the browser level.
+   * ```js
+   * fileOptions: [
+   *   {
+   *     exclude: "*",
+   *     include: "{versioned_directory}/*",
+   *     cacheControl: "public,max-age=31536000,immutable",
+   *   },
+   *   {
+   *     exclude: "*",
+   *     include: "[{non_versioned_file1}, {non_versioned_file2}, ...]",
+   *     cacheControl: "public,max-age=0,s-maxage=31536000,must-revalidate",
+   *   },
+   *   {
+   *     exclude: "*",
+   *     include: "[{non_versioned_dir_1}/*, {non_versioned_dir_2}/*, ...]",
+   *     cacheControl: "public,max-age=0,s-maxage=31536000,must-revalidate",
+   *   },
+   * ]
+   * ```
+   *
+   * @example
+   * ```js
+   * fileOptions: [
+   *   {
+   *     exclude: "*",
+   *     include: "{versioned_directory}/*.css",
+   *     cacheControl: "public,max-age=31536000,immutable",
+   *     contentType: "text/css; charset=UTF-8",
+   *   },
+   *   {
+   *     exclude: "*",
+   *     include: "{versioned_directory}/*.js",
+   *     cacheControl: "public,max-age=31536000,immutable",
+   *   },
+   *   {
+   *     exclude: "*",
+   *     include: "[{non_versioned_file1}, {non_versioned_file2}, ...]",
+   *     cacheControl: "public,max-age=0,s-maxage=31536000,must-revalidate",
+   *   },
+   *   {
+   *     exclude: "*",
+   *     include: "[{non_versioned_dir_1}/*, {non_versioned_dir_2}/*, ...]",
+   *     cacheControl: "public,max-age=0,s-maxage=31536000,must-revalidate",
+   *   },
+   * ]
+   * ```
+   */
+  fileOptions?: SsrSiteFileOptions[];
 }
 
 type SsrSiteNormalizedProps = SsrSiteProps & {
@@ -662,6 +718,8 @@ export abstract class SsrSite extends Construct implements SSTConstruct {
   }
 
   private createS3AssetFileOptions() {
+    if (this.props.fileOptions) return this.props.fileOptions;
+
     // Build file options
     const fileOptions = [];
     const clientPath = path.join(
@@ -670,7 +728,7 @@ export abstract class SsrSite extends Construct implements SSTConstruct {
     );
     for (const item of fs.readdirSync(clientPath)) {
       // Versioned files will be cached for 1 year (immutable) both at
-      // CDN and browser level.
+      // the CDN and browser level.
       if (item === this.buildConfig.clientBuildVersionedSubDir) {
         fileOptions.push({
           exclude: "*",
@@ -721,7 +779,7 @@ export abstract class SsrSite extends Construct implements SSTConstruct {
   private createS3Deployment(
     cliLayer: AwsCliLayer,
     assets: Asset[],
-    fileOptions: { exclude: string; include: string; cacheControl: string }[]
+    fileOptions: SsrSiteFileOptions[]
   ): CustomResource {
     // Create a Lambda function that will be doing the uploading
     const uploader = new CdkFunction(this, "S3Uploader", {
@@ -765,15 +823,19 @@ export abstract class SsrSite extends Construct implements SSTConstruct {
         })),
         DestinationBucketName: this.bucket.bucketName,
         FileOptions: (fileOptions || []).map(
-          ({ exclude, include, cacheControl }) => {
+          ({ exclude, include, cacheControl, contentType }) => {
+            if (typeof exclude === "string") {
+              exclude = [exclude];
+            }
+            if (typeof include === "string") {
+              include = [include];
+            }
             return [
-              "--exclude",
-              exclude,
-              "--include",
-              include,
-              "--cache-control",
-              cacheControl,
-            ];
+              ...exclude.map((per) => ["--exclude", per]),
+              ...include.map((per) => ["--include", per]),
+              ["--cache-control", cacheControl],
+              contentType ? ["--content-type", contentType] : [],
+            ].flat();
           }
         ),
         ReplaceValues: this.getS3ContentReplaceValues(),